# SegmentNodeList::getSplitCoordinates(): fix quadratic performance pattern