@dr-jts 2021-06-16
Good find.

I note that even with this change the GEOS code is [different](https://github.com/locationtech/jts/blob/master/modules/core/src/main/java/org/locationtech/jts/noding/SegmentNodeList.java#L281-L289) to JTS.  JTS uses the `CoordinateList` class, which prevents duplicate coordinates from being inserted (as opposed to removing them afterwards).   Would this approach be any faster than using `coordList.erase(std::unique(...))`?  If so, perhaps it's worth using the GEOS `CoordinateList` here.