@rouault 2021-06-16
> Would this approach be any faster than using coordList.erase(std::unique(...))? 

I anticipate it to be marginally / not measurably faster. I can give it a try though.

>  If so, perhaps it's worth using the GEOS `CoordinateList` here.

Side note: I think the port of CoordinateList to C++ isn't ideal by using a std::list wherea JTS use an ArrayList. In C++, a std::list<> is really a double-linked list, wheras ArrayList is mostly a vector. I read an article recently whose point was: you should hardly ever use std::list<>