@rouault 2021-06-16
> I anticipate it to be marginally / not measurably faster. I can give it a try though.

My anticipation was right. I've added a commit that checks if points are duplicates before adding. The timings are exactly the same as with the first commit