@pramsey 2021-06-16
If it's not faster, but it's way more invasive... ?