@Komzpa 2021-06-16
I confirm that in the slow buffer profiles I have on hand I see a lot of geos::noding::SegmentNodeList::addEdgeCoordinates. It also feels like it leaks like hell (32 processes of Postgres manage to get 700GB of RAM consumed even though input is guaranteed <1GB), but I have no good proofs.