@dr-jts 2021-06-16
> > If it's not faster, but it's way more invasive... ?
> 
> I'm happy to only keep the first commit

+1 from me.  Thanks for checking.