@dr-jts 2021-06-16
> Side note: I think the port of CoordinateList to C++ isn't ideal by using a std::list wherea JTS use an ArrayList. In C++, a std::list<> is really a double-linked list, wheras ArrayList is mostly a vector. I read an article recently whose point was: you should hardly ever use std::list<>

Good point!  That should definitely be changed.  Not sure how much GEOS `CoordinateList` is used - in JTS it's used quite a bit.