@rouault 2021-06-16
> > I'm happy to only keep the first commit
> 
> +1 from me. Thanks for checking.

Second commit dropped