@pramsey 2021-06-16
Also back-ported into 3.9